fix: minor fixes, prepare for client
